scraper/requirements.txt
beautifulsoup4==4.12.3
playwright>=1.40.0
supabase==2.3.0
httpx==0.24.1
httpcore==0.17.3
gotrue==2.5.0
postgrest==0.13.2
realtime==1.0.6
storage3==0.6.1
python-dotenv>=1.0.0
